# gh-118761: Improve import time for `csv`